requirements.txt
python-dotenv==1.0.1
cassandra-driver==3.28.0
locust==2.24.0
pyspark==3.5.0